Copybara import of Pigweed
Changes 84694379..a2482015 (some omitted):

  a2482015  docs: Switch to Pigweed doxylink mirror
  f0baae60  pw_async2: Fix Android.bp for internal yield header
  01adad65  pw_metric: Wrap metrics lists in custom type
  76ef269c  pw_protobuf: Expose the field key when decoding
  b49a2916  pw_async2: Categorize FutureTask as a task instead of a future in docs
  fb12a0f8  pw_async2: Clarify future completion semantics
  246a3638  pw_bytes: Fix implicit conversion warnings
  e164d2fa  pw_async2: Dispatcher stress test
  b78be75d  pw_ide: Improve Bazel path resolution with symlink prefix support
  bddecd80  pw_log: Fix default log level comment; migrate to Doxygen
  922e20b0  pw_build: Add deps handling to pw_linker_script
  b5ae97d6  pw_bluetooth_proxy: Use v1_adapter::MultiBuf
  f18e4d75  pw_containers: Optimize DynamicMap::erase(const_iterator)
  52604898  pw_kernel: Panic on double lock
  21432ee5  pw_kernel: Properly break scheduler lock on thread start
  7fc049d8  pw_kernel: Implement lock propagation in thread/process join signalling
  5aa5d5b4  pw_bluetooth_sapphire: Downgrade log for unacked search results
  5e2fab3d  pw_async2: Fix deregistering a task that is woken while running
  fd8bb48e  pw_async2: Do not hold the lock during Wake()
  f0ac35a1  pw_allocator: Assert ok result in Init()
  c91d97a6  pw_async2: Remove std::variant; use ReadyType instead of Empty
  6685f88d  pw_async2: Add BoxedFuture and Box helper
  16854787  pw_kernel: Don't compile tests into system images
  d8447066  pw_bluetooth_sapphire: Add LE connection Inspect data
  465ef5af  pw_bluetooth_sapphire: Add BR/EDR connection Inspect data
  8c58dbaf  pw_format: Remove unused unstable features
  5de921bb  pw_async2: Add default case to handling Context for CoroTask
  424cd784  pw_kernel: Increase wait_group stack sizes to match other tests
  53b7d529  pw_presubmit: Exclude roll changes from check
  934faee2  pw_protobuf: Remove in progress note from docs
  dc59fc1f  pw_async2: Use PW_AWAIT instead of TRY_READY*
  f080d75b  pw_async2: Update Join to move values out on completion
  db13df14  pw_async2: Switch Dispatcher to IntrusiveForwardList
  4f7f36d2  pw_bluetooth: Add btsnoop epoch delta
  60041564  pw_package: Log output when git clone fails
  381b193c  pw_toolchain: Expose toolchain tools and add docs
  8ffc64a4  pw_bluetooth_proxy: Fix build fail from FreeRTOS system
  c2a60fa4  pw_ide: Add a Local GEMINI.md for more specific guidance
  5a9beee1  pw_kernel: Preserve system template evaluation error context
  e21bd359  agents: Refactor code review workflow into a skill
  5513af6b  pw_containers: Merge CountAndCapacity into GenericDeque
  cb45bc3c  pw_bluetooth_proxy: Create Bazel aliases for gatt and rfcomm
  ea310a23  agents: Create bluetooth skill
  6bb3e490  pw_async2: Move coroutine examples to source file; JoinFuture fix
  58b7f381  pw_bluetooth: Add LEGetControllerActivityEnergyInfoCommand
  4f20d144  pw_bluetooth_sapphire: Add documentation on how to use local Fuchsia SDK
  327fbe0b  bazel: Delete WORKSPACE file
  e31fe496  pw_async2: coro.h Add missing <variant> include
  7eb2a4b1  pw_async2: Clean up Waker and Task
  ef2647df  multibuf: Remove SingleChunkRegionTracker memset on release
  ef1eee59  pw_ide: Introduce ide_query tool
  0fee24d8  pw_ide: Extract dependencies and relativize paths in compile commands
  c6efb633  pw_kernel: Transition processes to use an ownership model
  35f651ca  pw_async2: Simplify coroutine allocation
  86e48602  pw_async2: Pass CoroContext by value
  caee934d  pw_bluetooth_sapphire: Add LowEnergyConnectionHandle::TransferPASync
  ee0b087a  pw_env_setup: Roll pyperclip to 1.11.0
  d11fe72b  pw_async2: Rework task allocation
  1a542aa3  bazel: Improve the robustness of the cred_helper
  1cecd4d0  pw_toolchain: Declare rustfmt and rust_analyzer toolchains for non-host
  fb0b0cca  pw_work_queue: Add clear() to CustomWorkQueue
  992c1acc  pw_multibuf: Unify unit tests
  f5da5b2d  pw_async2: Add PW_AWAIT macro
  577355b0  pw_containers: Fix const correctness in iterators
  8e88d07a  pw_bluetooth_sapphire: Enable specifying allocator
  e6eb48d7  pw_bluetooth_sapphire: Add MultiBuf support to CommandChannel::OnEvent
  9f6e1d2b  pw_async2: Support implicit conversion from Allocator for CoroContext
  01eb34ee  pw_kernel: Add async IPC syscalls
  311158fc  pw_bluetooth_sapphire: Add checks for channel liveness in ProfileServer
  1b2b05d1  MODULE.bazel: Use the LLVM_VERSION variable
  3cb367a1  pw_async2: Minor channel docs improvements
  be77105f  pw_result: Add Android.bp dep on pw_containers
  45dfce49  pw_async2: Rework pw_async2 coroutines
  2c9165a1  pw_log_rpc: Allow on_open_callback retention
  e0dfe708  pw_kernel: Add process termination
  7180c1fc  pw_proto_compiler: Respect dir_pw_third_party_protobuf
  1cd56d38  rp2040: Remove building specific modules recommendation
  4dc4a90f  pw_containers: Add pw::DynamicMap
  21779c05  workflows: Add build_rp2xxx_tests to default
  ef387820  pw_bluetooth: Add sniff subrating event definition
  edfc8046  pw_bluetooth: Add Snoop Log Trace
  037f3859  pw_kernel: Add context switch tracing
  bdcf600a  pw_console: Fix invalid regex escape sequence
  005649ae  pw_kernel: Add PMSAv7 MPU validation to system generator
  32917062  pw_kernel: Don't compile in kernel userspace code in a kernel only app
  25af0019  pw_containers: Optional type with custom state enum
  694ae916  pw_result: Make internal StatusOrData shareable as OptionalData
  0ad0ca4d  bazel: Update pip.parse configuration
  44aad11f  pw_protobuf: Add common_nanopb target
  a8bdfa66  rust: Upgrade rules_rust
  137aa782  perfetto: Disable rust target for trace proto on MCU's
  58f58dd0  rust: Add zngur dependency to crates_std

GitOrigin-RevId: a2482015ded981d31dfc052d151234aac328f83e

+++ pw_bytes/bit_test.cc
@@ -25,15 +25,18 @@ namespace {
 
 // SignExtend unsigned integer values into signed integers.
 constexpr uint32_t kNegative24Bits = 0x00FACADE;
-constexpr int32_t kExtendedNegative24Bits = SignExtend<24>(kNegative24Bits);
+constexpr int32_t kExtendedNegative24Bits =
+    static_cast<int32_t>(SignExtend<24>(kNegative24Bits));
 static_assert(kExtendedNegative24Bits == static_cast<int32_t>(0xFFFACADE));
 
 constexpr uint32_t kPositive20Bits = 0x00000ACE;
-constexpr int32_t kExtendedPositive20Bits = SignExtend<20>(kPositive20Bits);
+constexpr int32_t kExtendedPositive20Bits =
+    static_cast<int32_t>(SignExtend<20>(kPositive20Bits));
 static_assert(kExtendedPositive20Bits == static_cast<int32_t>(0x00000ACE));
 
 constexpr uint32_t kNegative12Bits = 0x00000ACE;
-constexpr int32_t kExtendedNegative12Bits = SignExtend<12>(kNegative12Bits);
+constexpr int32_t kExtendedNegative12Bits =
+    static_cast<int32_t>(SignExtend<12>(kNegative12Bits));
 static_assert(kExtendedNegative12Bits == static_cast<int32_t>(0xFFFFFACE));
 
 constexpr uint32_t k32Bits = 0b10100000101000001010000010100000;
